refactor: constants 파일명 변경 및 불필요 코드 제거
